removing unused file
