+++ TipCalc/TipCalc/Localizables/Localizable.xcstrings
@@ -1,18 +1,6 @@
 {
   "sourceLanguage" : "en",
   "strings" : {
-    "-" : {
-
-    },
-    "$8.90" : {
-
-    },
-    "$89.00" : {
-
-    },
-    "$97.00" : {
-
-    },
     "Calculate tip" : {
       "comment" : "Calculate tip",
       "localizations" : {
@@ -57,9 +45,6 @@
         }
       }
     },
-    "Initial bill" : {
-
-    },
     "Input bill value" : {
       "comment" : "Input the bill value",
       "localizations" : {
@@ -70,9 +55,6 @@
           }
         }
       }
-    },
-    "Mar 4, 2025 at 1:100 PM" : {
-
     },
     "No value to be saved" : {
       "comment" : "No value to be saved!",
@@ -104,9 +86,6 @@
           }
         }
       }
-    },
-    "tip" : {
-
     },
     "Tip value" : {
       "comment" : "Tip value",
